requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
werkzeug==3.0.1

